+++ test/shared/fixtures.ts
@@ -1,0 +1,3 @@
+import { MockContract, smockit } from "@eth-optimism/smock"
+import { ethers } from "hardhat"
+